Hotbar prototype number 2 completed
Created the customer hotbar. This time, it follows the arch spec a little more closely, with a UIController with an updateUI method which the player may call when the customer UI needs to be updated. There are a lot of placeholders in this prototype. For example, the customers in my scene handle their own timers, but this will be handled by the player later on. Also, many fields are marked as Serializable for the sake of testing. This is not entirely reflective of the final product, but it should give a picture of what it might be. If you want to see another example for a hotbar, see the branch, Liam-Hotbar (the one I showed during last week's presentation).